+++ crates/nostr/CHANGELOG.md
@@ -82,6 +82,7 @@
 - Add `UnsignedEvent::mine` and `UnsignedEvent::mine_async` (https://github.com/rust-nostr/nostr/pull/1334)
 - Add `banner`, `self`, `terms_of_service` to `RelayInformationDocument` (https://github.com/rust-nostr/nostr/pull/1336)
 - Add `restricted_writes`, `default_limit` to `Limitation` (https://github.com/rust-nostr/nostr/pull/1336)
+- Add NIP-66 kinds and relay discovery tags (https://github.com/rust-nostr/nostr/pull/1346)
 
 ### Removed